Cbrelease 4.0.1 (#504)
* Disabled explore button before login flow

* Fixed update/add roles flow  issues

* user onboarding fixes

* Fixed roles&activity add issue

* Fixed lint issue

* time fix

* New API structure To play a Practice Assessment Updated

* New API structure To play a Practice Assessment Updated #1

+++ project/ws/app/src/lib/routes/profile-v3/components/topic-card/topic-card.component.ts
@@ -1,6 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core'
+import { MatDialog } from '@angular/material'
 // tslint:disable-next-line
 import _ from 'lodash'
+import { DialogConfirmComponent } from 'src/app/component/dialog-confirm/dialog-confirm.component'
 import { NSProfileDataV3 } from '../../models/profile-v3.models'
 import { TopicService } from '../../services/topics.service'
 
@@ -8,15 +10,15 @@ import { TopicService } from '../../services/topics.service'
   selector: 'ws-app-topic-card',
   templateUrl: './topic-card.component.html',
   styleUrls: ['./topic-card.component.scss'],
-    /* tslint:disable */
-    host: { class: 'flex flex-1 top_main flex-col' },
-    /* tslint:enable */
+  /* tslint:disable */
+  host: { class: 'flex flex-1 top_main flex-col' },
+  /* tslint:enable */
 })
 export class TopicCardComponent implements OnInit {
   @Input() topic!: NSProfileDataV3.ITopic
   show = 6
   // selectedTopics: Subscription | null = null
-  constructor(private topicService: TopicService) { }
+  constructor(private topicService: TopicService, private dialog: MatDialog) { }
 
   ngOnInit() {
 
@@ -28,6 +30,7 @@ export class TopicCardComponent implements OnInit {
       if (index !== -1) {
         /// remove from store
         this.topicService.removeSystemTopics(top)
+
       } else {
         /// add to store
         this.topicService.addSystemTopics(top)
@@ -37,7 +40,16 @@ export class TopicCardComponent implements OnInit {
       // const cIndex = _.indexOf(this.topicService.getCurrentSelectedTopics[index].children, top)
       if (index !== -1) {
         /// remove from store
-        this.topicService.removeDesiredTopics(top)
+        const data = { title: 'Confirmation Needed', body: 'Do you want to delete the Topic?' }
+        const confirmationDialog = this.dialog.open(DialogConfirmComponent, {
+          autoFocus: false,
+          // tslint:disable-next-line:object-shorthand-properties-first
+          data,
+        })
+        confirmationDialog.afterClosed().subscribe(result => {
+          // tslint:disable-next-line:block-spacing
+          if (result) {this.topicService.removeDesiredTopics(top) }
+        })
       } else {
         /// add to store
         this.topicService.addDesiredTopics(top)
@@ -65,5 +77,8 @@ export class TopicCardComponent implements OnInit {
   }
   showMore() {
     this.show += 10
+  }
+  showLess() {
+    this.show = 6
   }
 }
